requirements_working.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.2
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==1.40.0
networkx==3.2.1
numpy==2.0.2
openai==1.82.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pinecone==7.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
pluggy==1.6.0
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
zipp==3.21.0
zstandard==0.23.0
